3. Backend/order_mgmt_v1/requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
flask>=3.0.0
python-dotenv>=1.0.0
